chore: remove nanoid (#164)
## What

- Removes `nanoid` from our dependencies
- Moves idempotencyKey setting outside the retry loop
- Removes `idempotencyKey` from the server's schema

## Why

- This is a quickstart example, meant to be as copy-pasteable as
possible. Every dependency introduces additional complexity that 
the user may not need / want. Nanoid is now ESM only on version 4.x - 
rather than lock it to 3.x I think removing the dependency (and any other
superfluous dependencies) is the better path forward.
- I noticed in passing that we are setting the idempotencyKey on every
retry, but considering when this code runs, setting an idempotencyKey _inside_
the retry will cause a new key to be created on every retry. We should have
a single idempotencyKey set for the request across retries.
- Finally, PR review noticed that payload.idempotencyKey is outdated and
not necessary anymore, so we removed it from the schema.